Add custom exceptions and response

+++ src/PhenGlobal.EmployeeService.Application/Features/LeaveTypes/Commands/DeleteLeaveType/DeleteLeaveTypeCommandHandler.cs
@@ -1,5 +1,7 @@
 using MediatR;
 using PhenGlobal.EmployeeService.Application.Persistence.Contracts;
+using PhenGlobal.EmployeeService.Application.Exceptions;
+using PhenGlobal.EmployeeService.Domain.Entities;
 
 namespace PhenGlobal.EmployeeService.Application.Features.LeaveTypes.Commands.DeleteLeaveType
 {
@@ -15,6 +17,11 @@ namespace PhenGlobal.EmployeeService.Application.Features.LeaveTypes.Commands.De
         public async Task<Unit> Handle(DeleteLeaveTypeCommand request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+
+            if(leaveType is null) {
+                throw new NotFoundException(nameof(LeaveType), request.Id);
+            }
+
             await _leaveTypeRepository.Delete(leaveType);
             return Unit.Value;
         }
